+++ frontend/src/components/Layout.jsx
@@ -1,6 +1,5 @@
 import { Outlet } from 'react-router-dom'
 import Sidebar from './Sidebar'
-import Header from './Header'
 import Player from './Player'
 import ToastContainer from './ToastContainer'
 
@@ -10,7 +9,6 @@ export default function Layout() {
       <Sidebar />
 
       <main className="main-content">
-        <Header />
         <div className="page-content">
           <Outlet />
         </div>